+++ src/db/redis_common.js
@@ -1,5 +1,27 @@
-import * as cacheDb from './cache/redis';
-import * as longTermDb from './long_term/redis';
+/**
+ * Copyright (c) 2018, 2019 National Digital ID COMPANY LIMITED
+ *
+ * This file is part of NDID software.
+ *
+ * NDID is the free software: you can redistribute it and/or modify it under
+ * the terms of the Affero GNU General Public License as published by the
+ * Free Software Foundation, either version 3 of the License, or any later
+ * version.
+ *
+ * NDID is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the Affero GNU General Public License for more details.
+ *
+ * You should have received a copy of the Affero GNU General Public License
+ * along with the NDID source code. If not, see https://www.gnu.org/licenses/agpl.txt.
+ *
+ * Please contact info@ndid.co.th for any further questions
+ *
+ */
+
+import cacheDbRedisInstance from './cache/redis';
+import longTermDbRedisInstance from './long_term/redis';
 
 import CustomError from '../error/custom_error';
 import errorType from '../error/type';
@@ -7,9 +29,9 @@ import errorType from '../error/type';
 function getRedis(dbName) {
   switch (dbName) {
     case 'cache':
-      return cacheDb.redis;
+      return cacheDbRedisInstance.redis;
     case 'long-term':
-      return longTermDb.redis;
+      return longTermDbRedisInstance.redis;
     default:
       throw new CustomError({ message: 'Unknown database name' });
   }